Add all supplier related classes and attributes into a supplier package
